Various fixes I didn't push from January.  - Fix build 17 form config.  - When there's no known WAD file, use the theme specified in the Frogger map.  - Create spline math representations shared between games.  - Allow highlighting entities when clicked in old Frogger.  - Fix CVector editors to allow proper preview and editing.  - Finish editor for difficulty data in old Frogger.  - Improve logger specification for old Frogger.  - Finished real-time gouraud shading editor, even if it's mega slow.  - Improve old Frogger texture flipping and mesh updates.  - Show old Frogger map camera offset at the right world position.  - Generalize the 2D selection rectangle seen in the camera height-field editor.  - Create UISidePanel abstraction.  - Add polygon editor for old Frogger.  - Updating vertices now updates the map mesh in real-time.  - Fix some scenarios with inaccurate gouraud shading recreations.  - Add MIT License  - Add translation gizmos to the MeshMarker system.  - Properly handle key presses in 3D when multiple pieces of code might want to know about the same key.  - Generalize axis plane creation for gizmos.  - Shared features for updating mesh entries across node & mesh definitions.  - Reduced overhead of texture atlas rebuilds.  - Fixed a bug causing texture sheets to avoid using large amounts of space when a new largest texture is added.

+++ src/net/highwayfrogs/editor/file/standard/SVector.java
@@ -274,8 +274,8 @@ public class SVector extends GameObject implements Vector {
      * @return distanceSquared
      */
     public double distanceSquared(SVector other) {
-        return ((other.getFloatX() - getFloatX()) * (other.getFloatX() - getFloatX()))
-                + ((other.getFloatY() - getFloatY()) * (other.getFloatY() - getFloatY()))
-                + ((other.getFloatZ() - getFloatZ()) * (other.getFloatZ() - getFloatZ()));
+        return ((double) (other.getFloatX() - getFloatX()) * (double) (other.getFloatX() - getFloatX()))
+                + ((double) (other.getFloatY() - getFloatY()) * (double) (other.getFloatY() - getFloatY()))
+                + ((double) (other.getFloatZ() - getFloatZ()) * (double) (other.getFloatZ() - getFloatZ()));
     }
 }
